refactor: 전략 상세페이지 타이틀, 헤더 컴포넌트 분리 (#94)
